grafana: preload datasource and dashboards
Grafana has a concept of provisioning, which allows you to define
datasources that should be loaded when grafana starts up. Additionally,
you can also preload dashboards.

The only downside is that the dashboards must hard-code the datasource
name instead of being able to use an input variable and templating
(which is possible when doing import/export via the UI. This is a
current limitation in grafana and there's an open issue #10786)

Given that we are providing these dashboards for Gimbal, and the fact
that we provide a prometheus deployment out of the box, hard-coding the
datasource name does not seem like a problem.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>