discovery: update translators to use new naming conventions
Instead of appending the cluster name to the service name, the
discovered name will begin with the cluster name, and have the backend
service name appended. This ensures the discovered name begins with a
letter, as required by the Kubernetes DNS_LABEL spec.

On the OpenStack side, labels that are added to the discovered
services/endpoints are also shortened in the case that they go over the
character limit.

Additionally, the shortening process was updated to use the hash of
each component, instead of the hash of the discovered name as a whole.
The reason for this is that services that have long names on different
backends will end up having the same hash.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>