discovery: fix syncqueue retries
The client-go workqueue expects the user to mark any item that needs to
be retried as "dirty". That is, on failure, the item must be re-added to
the queue for reprocessing at a later time.

The item is added to the queue via AddRateLimited, which will add the
item once the rate limiter says it's okay.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>